Enable dvc to break down data processing into stages to have data lineage
