--- .gitignore
@@ -1,3 +1,0 @@
-.envrc
-.idea
-__pycache__
