+++ wolfram_alpha_api/data/.gitignore
@@ -1,0 +1,2 @@
+/retrieve
+/extract
