+++ wolfram_alpha_api/envrc.template
@@ -1,0 +1,2 @@
+export WOLFRAM_APP_ID=""
+export WOLFRAM_URL="https://api.wolframalpha.com/v2/query"
