+++ wolfram_alpha_api/src/common/pyproject.toml
@@ -1,0 +1,19 @@
+[project]
+name = "common"
+version = "0.1.0"
+description = ""
+authors = [
+    {name = "Darshan Kothari",email = "dkothari777@gmail.com"}
+]
+requires-python = ">=3.8"
+
+[tool.dev.dependencies]
+
+
+[tool.poetry]
+packages = [{include = "common"}]
+
+
+[build-system]
+requires = ["poetry-core>=2.0.0,<3.0.0"]
+build-backend = "poetry.core.masonry.api"
