+++ wolfram_alpha_api/src/retrieve/pyproject.toml
@@ -1,0 +1,19 @@
+[project]
+name = "retrieve"
+version = "0.1.0"
+description = ""
+authors = [
+    {name = "Darshan Kothari",email = "dkothari777@gmail.com"}
+]
+requires-python = ">=3.8"
+
+[tool.dev.dependencies]
+common = { path = "../common/", develop = true}
+
+[tool.poetry]
+packages = [{include = "retrieve"}]
+
+
+[build-system]
+requires = ["poetry-core>=2.0.0,<3.0.0"]
+build-backend = "poetry.core.masonry.api"
